Update example to use u32 instead of usize

+++ adder/src/lib.rs
@@ -1,5 +1,5 @@
 #[no_mangle]
-pub extern "C" fn add(left: usize, right: usize) -> usize {
+pub extern "C" fn add(left: u32, right: u32) -> u32 {
     left + right
 }
